ci skip: dbg header

+++ client.go
@@ -55,9 +55,9 @@ func (c *Client) Now() (Now, error) {
 
 	size := binary.LittleEndian.Uint32(c.m[8:12])
 	log.Printf("size: %d\n", size)
-	ver := binary.LittleEndian.Uint32(c.m[12:14])
+	ver := binary.LittleEndian.Uint16(c.m[12:14])
 	log.Printf("version: %d\n", ver)
-	gen := binary.LittleEndian.Uint32(c.m[14:16])
+	gen := binary.LittleEndian.Uint16(c.m[14:16])
 	log.Printf("generation: %d\n", gen)
 
 	asof_s := binary.LittleEndian.Uint64(c.m[16:24])
